[NEW]     Added OMImageFlow (A coverflow style control for images) [NEW]     Added checked property to OMButton along with checked image [NEW]     Added full 3D rendering in OpenMobile [NEW]     Added class GradientData to allow for custom gradient effects on rectangles and images [NEW]     Added SupportedOSConfigurations attribute to enable plugins to set the OS configurations flags that's required           to use the plugin. [NEW]     Added support for disabling / enabling plugins via the settings plugin [NEW]     Started working on using mPlayer as a media player [NEW]     Started work on new iMediaProvider interface that will replace the current iAVPlayer and iTunedContent interfaces [CHANGE]  Changed SystemEvent delegate to a more flexible format using an object array as arguments [NEW]     Added control OMTargetWindow which is used to get a valid window handle [NEW]     Added autofit to OMLabels and any controls deriving from label [BUGFIX]  Updated OMClock to center clock in relation to the controls area

+++ FrontEnd/OpenMobile.Framework/Controls/OMGauge.cs
@@ -910,7 +910,7 @@ namespace OpenMobile.Controls
 
             //Draw Digital Value
             Rectangle digiFRect = new Rectangle(this.Left + this.Width / 2 - this.width / 7, this.Top + (int)(this.height / 1.18), this.width / 4, this.Height / 12);
-            g.FillRectangle(new Brush(Color.FromArgb(30, Color.Gray)), (this.Left + (float)this.Width / 2F - (float)this.width / 5F), (this.Top + (float)this.height / 1.2F), ((float)this.width / 2.5F), ((float)this.Height / 9F));
+            g.FillRectangle(new Brush(Color.FromArgb(30, Color.Gray)), (int)((this.Left + (float)this.Width / 2F - (float)this.width / 5F)), (int)((this.Top + (float)this.height / 1.2F)), (int)((float)this.width / 2.5F), (int)((float)this.Height / 9F));
             DisplayNumber(g, this.currentValue, digiFRect);
 
             //Paint the needle
